requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
gunicorn==20.1.0
requests==2.31.0
Werkzeug==2.3.3
PyJWT==2.3.0
pytest==8.3.3